ping-google/requirements.txt
requests
googlesearch-python 